Manager Final

+++ resources/views/followups/index.blade.php
@@ -26,12 +26,21 @@
                 <div class="card-title">
                   <div class="row">
                      <div class="col col-sm-8">
+                     <form method="GET" action="/followups" class="form-inline" style="width: 100%;">
+                          <select class="form-control" name="employee" style="margin-right: 4px;width: 200px;">
+                            <option>Rahul</option>
+                            <option>Irfan</option>
+                            <option>Amit</option>
+                            <option>Ashish</option>
+                          </select>
+                          <button type="submit" class="btn btn-success">Submit</button>
+                        </form>
                         
                      </div>
                      <div class="col col-sm-4">
                         <form method="GET" action="/followups" class="form-inline float-right">
                           <input type="date" value="{{ $date }}" name="date" id="date" class="form-control" style="margin-right: 4px;">
-                          <button type="submit" class="btn btn-primary">Submit</button>
+                          <button type="submit" class="btn btn-success">Submit</button>
                         </form>
                      </div>
                   </div>   
@@ -46,7 +55,9 @@
                           <th>Phone</th>
                           <th>Enquiry Date</th>
                           <th>Model</th>
+                           @if(auth()->user()->isAdmin())
                           <th>Actions</th>
+                          @endif
                         </tr>
                       </thead>
                       <tbody>
@@ -59,16 +70,21 @@
                               <td>{{ $enquiry->phone }}</td>
                               <td>{{ $enquiry->created_at->format('d-m-Y') }}</td>
                               <td>{{ $enquiry->vehicle->name }}</td>
-                              <td><a href="{{ route('enquiries.show', ['enquiry' => $enquiry->id ])}}" class="btn btn-primary btn-sm">View</a> 
-                                  <a href="{{ route('enquiries.edit', ['enquiry' => $enquiry->id ])}}" class="btn btn-success btn-sm">Edit</a> 
-                                  <a href="#" @click="deleteEnquiry({{ $enquiry->id  }})" class="btn btn-danger btn-sm">Delete</a>
-                              </td>
+                              @if(auth()->user()->isAdmin())
+                                <td>
+                                     <a href="{{ route('enquiries.show', ['enquiry' => $enquiry->id ])}}" class="btn btn-primary btn-sm">View</a> 
+                                    <a href="{{ route('enquiries.edit', ['enquiry' => $enquiry->id ])}}" class="btn btn-success btn-sm">Edit</a> 
+                                    <a href="#" @click="deleteEnquiry({{ $enquiry->id  }})" class="btn btn-danger btn-sm">Delete</a>
+                                </td>
+                              @endif
                             </tr>
 
                         @endforeach
                         
                       </tbody>
                     </table>
+
+                    <a href="#" class="btn btn-primary">Print List</a>
                 </div>
             </div>
